move MatrixVT to layers
